+++ Source/Krypton Components/Krypton.Ribbon/View Layout/ViewLayoutRibbonQATFromOverflow.cs
@@ -59,8 +59,7 @@ internal class ViewLayoutRibbonQATFromOverflow : ViewLayoutRibbonQATContents
             var qatOverflow = new List<IQuickAccessToolbarButton>();
 
             // Scan all the defined buttons for ones to show as overflowing
-            foreach (IQuickAccessToolbarButton qatButton in from IQuickAccessToolbarButton qatButton in Ribbon.QATButtons
-                                                            select qatButton)
+            foreach (var qatButton in Ribbon.QATButtons.OfType<IQuickAccessToolbarButton>())
             {
                 // If the button requests to be shown...
                 if (qatButton.GetVisible())