+++ Source/Krypton Components/Krypton.Ribbon/View Layout/ViewLayoutRibbonQATMini.cs
@@ -237,9 +237,8 @@ internal class ViewLayoutRibbonQATMini : ViewLayoutDocker
         Debug.Assert(context != null);
 
         // Scan to see if there are any visible quick access toolbar buttons
-        var visibleQATButtons = (from IQuickAccessToolbarButton qatButton in _ribbon.QATButtons
-                                 where qatButton.GetVisible()
-                                 select qatButton).Any();
+        var visibleQATButtons = _ribbon.QATButtons.OfType<IQuickAccessToolbarButton>()
+            .Any(static qatButton => qatButton.GetVisible());
             
         // Only show the border if there are some visible contents
         _border.Visible = visibleQATButtons;